requirements.txt
python-dotenv>=1.0.0
httpx>=0.27.0
websockets>=12.0
pandas>=2.2.0
numpy>=1.26.0
python-telegram-bot>=21.0
aiosqlite>=0.20.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
loguru>=0.7.2
pytz>=2024.1
pytest>=8.0.0
pytest-asyncio>=0.23.0
